Refactored common code
