+++ samcli/commands/list/resources/resources_context.py
@@ -3,29 +3,25 @@ Display the Resources of a SAM stack
 """
 import logging
 from typing import Optional
-import boto3
 
-from samcli.commands.exceptions import RegionError
-from samcli.lib.utils.boto_utils import get_boto_client_provider_with_config
-
+from samcli.commands.list.cli_common.list_common_context import ListContext
 from samcli.lib.list.resources.resource_mapping_producer import ResourceMappingProducer
 from samcli.lib.list.mapper_consumer_factory import MapperConsumerFactory
 from samcli.lib.list.list_interfaces import ProducersEnum
 
-
 LOG = logging.getLogger(__name__)
 
 
-class ResourcesContext:
+class ResourcesContext(ListContext):
     def __init__(
         self, stack_name: str, output: str, region: Optional[str], profile: Optional[str], template_file: Optional[str]
     ):
+        super().__init__()
         self.stack_name = stack_name
         self.output = output
         self.region = region
         self.profile = profile
         self.template_file = template_file
-        self.cloudformation_client = None
         self.iam_client = None
 
     def __enter__(self):
@@ -39,20 +35,8 @@ class ResourcesContext:
         """
         Initialize the clients being used by sam list.
         """
-        if not self.region:
-            session = boto3.Session()
-            region = session.region_name
-            if region:
-                self.region = region
-            else:
-                raise RegionError(
-                    message="No region was specified/found. "
-                    "Please provide a region via the --region parameter or by the AWS_REGION environment variable."
-                )
-
-        client_provider = get_boto_client_provider_with_config(region=self.region, profile=self.profile)
-        self.cloudformation_client = client_provider("cloudformation")
-        self.iam_client = client_provider("iam")
+        super().init_clients()
+        self.iam_client = self.client_provider("iam")
 
     def run(self) -> None:
         """
